refactor(logic): split work-phase handlers into focused modules
Decompose work-order application into dedicated handler files for purchase-land, explore, prospect, fixed-duration, and standard build targets so runWorkPhase remains a thin dispatcher and handler logic is independently testable. Refs #1958.

Made-with: Cursor
