+++ packages/colonizethis_logic/lib/src/orders/work_handlers/shared_work_assignment.dart
@@ -1,0 +1,90 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_logic/package_logger.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../constants.dart';
+
+final _log = packageLogger();
+
+void completeInstantCivilianOrder(
+  void Function(String, Unit) updateUnit,
+  Unit unit,
+  String targetTileKey,
+) {
+  updateUnit(
+    unit.id,
+    unit.copyWith(
+      status: UnitStatus.idle,
+      tileKey: targetTileKey,
+      clearOriginTileKey: true,
+      clearAssignedTileKey: true,
+      clearCurrentWork: true,
+    ),
+  );
+}
+
+bool tryAssignFixedDurationWorkOrder({
+  required WorkOrder order,
+  required Unit unit,
+  required String targetTileKey,
+  required String target,
+  required int totalTurns,
+  required int remainingTurns,
+  required void Function(String, Unit) updateUnit,
+}) {
+  if (!isWorkOrderTargetAllowedForUnitType(unit.type, target)) return false;
+  if (unit.currentWork != null || targetTileKey.isEmpty) return false;
+  _log.d(
+    'work order accepted and assigned unit=${order.unitId} target=$target targetTileKey=$targetTileKey totalTurns=$totalTurns',
+  );
+  updateUnit(
+    order.unitId,
+    unit.copyWith(
+      status: UnitStatus.working,
+      tileKey: targetTileKey,
+      originTileKey: unit.originTileKey ?? unit.tileKey,
+      assignedTileKey: targetTileKey,
+      currentWork: CurrentWork(
+        workTarget: target,
+        tileKey: targetTileKey,
+        totalTurns: totalTurns,
+        remainingTurns: remainingTurns,
+      ),
+    ),
+  );
+  return true;
+}
+
+bool tryAssignStealTechWorkOrder({
+  required WorkOrder order,
+  required Unit unit,
+  required String targetTileKey,
+  required void Function(String, Unit) updateUnit,
+}) {
+  return tryAssignFixedDurationWorkOrder(
+    order: order,
+    unit: unit,
+    targetTileKey: targetTileKey,
+    target: kWorkTargetStealTech,
+    totalTurns: totalTurnsForWork(kWorkTargetStealTech),
+    remainingTurns: totalTurnsForWork(kWorkTargetStealTech),
+    updateUnit: updateUnit,
+  );
+}
+
+bool tryAssignCounterSpyWorkOrder({
+  required WorkOrder order,
+  required Unit unit,
+  required String targetTileKey,
+  required void Function(String, Unit) updateUnit,
+}) {
+  return tryAssignFixedDurationWorkOrder(
+    order: order,
+    unit: unit,
+    targetTileKey: targetTileKey,
+    target: kWorkTargetCounterSpy,
+    totalTurns: totalTurnsForWork(kWorkTargetCounterSpy),
+    remainingTurns: 1,
+    updateUnit: updateUnit,
+  );
+}
